add date start icon and move icons to subcomponent folder
