+++ packages/date-picker/src/_subcomponents/Icons/IconLeft.tsx
