+++ packages/date-picker/src/_subcomponents/Icons/IconRight.tsx
