+++ packages/date-picker/src/_subcomponents/Icons/index.ts
@@ -1,0 +1,3 @@
+export * from "./IconDateStart"
+export * from "./IconLeft"
+export * from "./IconRight"
